requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
brave-search==0.1.8
certifi==2024.2.2
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
load-dotenv==0.1.0
multidict==6.0.5
numpy==1.26.4
openai==1.14.2
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
pytest-asyncio==0.23.6
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
